Feature supply management (#60)
* CRUD + Edit ContainerType Post Discount

* CREATE UPDATE Inbound, BillOfLading And Container

* CRUD Inbound

* CRUD outbound inbound

* CRUD Container

* FIX preAuthen Port ContainerType FIX check duplicate container

* FIX check

* UPDATE EDIT GET Booking and BillOfLading

* FIX Name function

* FIX Name function

* FIX Logic duplicate time when CRUD Inbound and billOfLading

* FIX dupplicate time when edit inbound

* FIX logic check create container driver in same time in container billoflading inbound

* FIX pickupTime must be before present ADD string.isEmpty()

* ADD isEmpty for outbound

* FIX isEmpty ADD @Transactional

* CHECK forwarder own driver

* CREATE function getInboundsByOutboundAndForwarder and check status equal combined

* ADD function GetContainerByBid

* CREATE function findForwarderByOutbound check time set emptyTime = pickupTime + 1 day

* FIX MODERATOR Auth

* ADD and CHANGE column in Outbound, SEARCH OutboundByStatus, FIX Double in edit function to String in discount and containerType

* GET forwarderId and merchantId by SecurityContextHolder when create

* CHECK Bidding

+++ src/main/java/com/crm/services/impl/ContainerServiceImpl.java
@@ -137,6 +137,12 @@ public class ContainerServiceImpl implements ContainerService {
     Container container = containerRepository.findById(request.getId())
         .orElseThrow(() -> new NotFoundException("ERROR: Container is not found."));
 
+    if (container.getStatus().equalsIgnoreCase(EnumSupplyStatus.COMBINED.name())
+        || container.getStatus().equalsIgnoreCase(EnumSupplyStatus.BIDDING.name())) {
+      throw new InternalException(
+          String.format("Container %s has been %s", container.getContainerNumber(), container.getStatus()));
+    }
+
     BillOfLading billOfLading = (BillOfLading) container.getBillOfLading();
 
     Set<Container> containers = billOfLading.getContainers();
@@ -213,8 +219,10 @@ public class ContainerServiceImpl implements ContainerService {
   public void removeContainer(Long id) {
     Container container = containerRepository.findById(id)
         .orElseThrow(() -> new NotFoundException("ERROR: Container is not found."));
-    if (container.getStatus().equalsIgnoreCase(EnumSupplyStatus.COMBINED.name())) {
-      throw new InternalException(String.format("Container %s has been combined", container.getContainerNumber()));
+    if (container.getStatus().equalsIgnoreCase(EnumSupplyStatus.COMBINED.name())
+        || container.getStatus().equalsIgnoreCase(EnumSupplyStatus.BIDDING.name())) {
+      throw new InternalException(
+          String.format("Container %s has been %s", container.getContainerNumber(), container.getStatus()));
     }
     containerRepository.delete(container);
   }
@@ -224,8 +232,10 @@ public class ContainerServiceImpl implements ContainerService {
 
     Container container = containerRepository.findById(id)
         .orElseThrow(() -> new NotFoundException("ERROR: Container is not found."));
-    if (container.getStatus().equalsIgnoreCase(EnumSupplyStatus.COMBINED.name())) {
-      throw new InternalException(String.format("Container %s has been combined", container.getContainerNumber()));
+    if (container.getStatus().equalsIgnoreCase(EnumSupplyStatus.COMBINED.name())
+        || container.getStatus().equalsIgnoreCase(EnumSupplyStatus.BIDDING.name())) {
+      throw new InternalException(
+          String.format("Container %s has been %s", container.getContainerNumber(), container.getStatus()));
     }
 
     BillOfLading billOfLading = (BillOfLading) container.getBillOfLading();
